refactor: Prep for page-specific layouts
