+++ src/pages/_app.tsx
@@ -1,172 +1,29 @@
 // Modules
-import { AnimatePresence, motion } from "framer-motion";
-
+import { NextPage } from "next";
 import type { AppProps } from "next/app";
-import Link from "next/link";
-import { useRouter } from "next/router";
 
-import { appWithTranslation, useTranslation } from "next-i18next";
+import { appWithTranslation } from "next-i18next";
 
-// SK Components
-import { MaterialIcon, PageLayout } from "@suankularb-components/react";
+import { ComponentType, ReactElement, ReactNode } from "react";
 
 // Styles
 import "@styles/global.css";
 
-// Animations
-import { fromUpToDown } from "@utils/animations/slide";
+// Components
+import Layout from "@components/Layout";
 
-const App = ({ Component, pageProps }: AppProps) => {
-  const router = useRouter();
-  const { t } = useTranslation();
+const App = ({
+  Component,
+  pageProps,
+}: AppProps & {
+  Component: NextPage & {
+    getLayout?: (page: ReactElement) => ReactNode;
+  };
+}) => {
+  // Use the layout defined at the page level, if available
+  const getLayout = Component.getLayout ?? ((page) => <Layout>{page}</Layout>);
 
-  // TODO: When logging in does become a thing, change this to a more sane implementation
-  const navItems = ["/", "/account/login", "/about"].includes(router.asPath)
-    ? [
-        {
-          name: t("navigation.home"),
-          icon: {
-            inactive: <MaterialIcon icon="home" type="outlined" />,
-            active: <MaterialIcon icon="home" type="filled" />,
-          },
-          url: "/",
-        },
-        {
-          name: t("navigation.login"),
-          icon: {
-            inactive: <MaterialIcon icon="login" type="outlined" />,
-            active: <MaterialIcon icon="login" type="filled" />,
-          },
-          url: "/account/login",
-        },
-        {
-          name: t("navigation.about"),
-          icon: {
-            inactive: <MaterialIcon icon="information" type="outlined" />,
-            active: <MaterialIcon icon="information" type="filled" />,
-          },
-          url: "/about",
-        },
-      ]
-    : [
-        "/s/home",
-        "/s/405/schedule",
-        "/s/405/class",
-        "/s/teachers",
-        "news",
-      ].includes(router.asPath)
-    ? [
-        {
-          name: t("navigation.home"),
-          icon: {
-            inactive: <MaterialIcon icon="home" type="outlined" />,
-            active: <MaterialIcon icon="home" type="filled" />,
-          },
-          url: "/s/home",
-        },
-        {
-          name: t("navigation.schedule"),
-          icon: {
-            inactive: <MaterialIcon icon="dashboard" type="outlined" />,
-            active: <MaterialIcon icon="dashboard" type="filled" />,
-          },
-          url: "/s/405/schedule",
-        },
-        {
-          name: t("navigation.class"),
-          icon: {
-            inactive: <MaterialIcon icon="groups" type="outlined" />,
-            active: <MaterialIcon icon="groups" type="filled" />,
-          },
-          url: "/s/405/class",
-        },
-        {
-          name: t("navigation.teachers"),
-          icon: {
-            inactive: <MaterialIcon icon="school" type="outlined" />,
-            active: <MaterialIcon icon="school" type="filled" />,
-          },
-          url: "/s/teachers",
-        },
-        {
-          name: t("navigation.news"),
-          icon: {
-            inactive: <MaterialIcon icon="newspaper" type="outlined" />,
-            active: <MaterialIcon icon="newspaper" type="filled" />,
-          },
-          url: "/news",
-        },
-      ]
-    : [
-        {
-          name: t("navigation.home"),
-          icon: {
-            inactive: <MaterialIcon icon="home" type="outlined" />,
-            active: <MaterialIcon icon="home" type="filled" />,
-          },
-          url: "/t/home",
-        },
-        {
-          name: t("navigation.schedule"),
-          icon: {
-            inactive: <MaterialIcon icon="dashboard" type="outlined" />,
-            active: <MaterialIcon icon="dashboard" type="filled" />,
-          },
-          url: "/t/schedule",
-        },
-        {
-          name: t("navigation.subjects"),
-          icon: {
-            inactive: <MaterialIcon icon="school" type="outlined" />,
-            active: <MaterialIcon icon="school" type="filled" />,
-          },
-          url: "/t/subjects/teaching",
-        },
-        {
-          name: t("navigation.class"),
-          icon: {
-            inactive: <MaterialIcon icon="groups" type="outlined" />,
-            active: <MaterialIcon icon="groups" type="filled" />,
-          },
-          url: "/t/509/class",
-        },
-        {
-          name: t("navigation.admin"),
-          icon: {
-            inactive: <MaterialIcon icon="security" type="outlined" />,
-            active: <MaterialIcon icon="security" type="filled" />,
-          },
-          url: "/t/admin",
-        },
-      ];
-
-  return (
-    <AnimatePresence
-      exitBeforeEnter
-      initial={false}
-      onExitComplete={() => window.scrollTo(0, 0)}
-    >
-      <div className="overflow-hidden bg-background">
-        <PageLayout
-          key={router.route}
-          currentPath={router.asPath}
-          navItems={navItems}
-          LinkElement={Link}
-        >
-          <motion.div
-            initial="hidden"
-            animate="enter"
-            exit="exit"
-            variants={fromUpToDown}
-            transition={{ duration: 0.25, ease: "easeInOut" }}
-            className="flex flex-grow flex-col overflow-auto"
-          >
-            <Component {...pageProps} />
-          </motion.div>
-        </PageLayout>
-      </div>
-    </AnimatePresence>
-  );
+  return getLayout(<Component {...pageProps} />);
 };
 
-export default appWithTranslation(App);
+export default appWithTranslation(App as ComponentType<AppProps>);
